end point tested with postman

+++ lms-dal/src/main/java/com/lms/dal/repository/BookRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface BookRepository extends PagingAndSortingRepository<EBook, Long> {
+public interface BookRepository extends PagingAndSortingRepository<EBook, Integer> {
 
     List<EBook> getBookByAuthor(EAuthor author);
