+++ lms-dal/src/main/java/com/lms/dal/repository/MemberRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MemberRepository extends PagingAndSortingRepository<EMember, Long> {
+public interface MemberRepository extends PagingAndSortingRepository<EMember, Integer> {
 
 }
